Данные берутся из слоя data, данные распределены по слоям по чистой архитектуре

+++ MyApplication/domain/src/main/java/com/example/domain/Hints/model/Checklist.kt
@@ -2,7 +2,7 @@ package com.example.domain.Hints.model
 
 //краткая инструкция в конце статьи (как краткий итог)
 data class Checklist (
-    val id: Long,
-    val articleId: Long,
+    val id: Int,
+    val articleId: Int,
     val items: List<String>
 )
